# go_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "backend",
        "contact": {},
        "license": {
            "name": "undamoni"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/detail": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get detail of previously traced URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sha256(URL)",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ResponseErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ResponseErr"
                        }
                    }
                }
            }
        },
        "/thumbnail": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get thumbnail of previously traced URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sha256(URL)",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "size of thumbnail(square, 100 \u003c size \u003c 800)",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ResponseErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ResponseErr"
                        }
                    }
                }
            }
        },
        "/trace": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Trace URL from safe machines",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL to trace",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ResponseErr"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ResponseErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.Result": {
            "type": "object",
            "properties": {
                "chains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "from_url": {
                    "type": "string"
                },
                "info": {
                    "$ref": "#/definitions/handlers.SiteInfo"
                },
                "term_url": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                }
            }
        },
        "handlers.SiteInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "main.ResponseErr": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}